feat(move-package-alt): Add first lockfile test (#8950)
# Description of change

This PR adds a manifest error type and support for codespan reporting
via `serde_spanned`. 
It also adds the first two simple validations for
package name and editions.

## Links to any relevant issues

Fixes #8336 .

## How the change has been tested

`cargo test` for move-package-alt
