+++ external-crates/move/crates/move-package-alt/src/dependency/git.rs
@@ -33,6 +33,7 @@ pub struct GitDependency<P = Unpinned> {
     /// The git commit-ish for the dep; guaranteed to be a commit if [P] is [Pinned].
     #[serde(rename = "git")]
     repo: String,
+
     rev: String,
 
     /// The path within the repository
