+++ external-crates/move/crates/move-package-alt/src/package/lockfile.rs
@@ -60,7 +60,9 @@ struct UnpublishedTable<F: MoveFlavor + fmt::Debug> {
 #[derive_where(Default, Clone)]
 #[serde(bound = "")]
 struct UnpublishedDependencies<F: MoveFlavor + fmt::Debug> {
+    #[serde(default)]
     pinned: BTreeMap<PackageName, PinnedDependencyInfo<F>>,
+    #[serde(default)]
     unpinned: BTreeMap<PackageName, ManifestDependencyInfo<F>>,
 }
