+++ external-crates/move/crates/move-package-alt/tests/.gitignore
@@ -1,0 +1,2 @@
+*.snap.new
+!**/*.lock
